perceptron.hh: 	Changed globalHistory in struct BPHistory from unsigned int to uint64_t

+++ src/cpu/pred/perceptron.hh
@@ -31,7 +31,7 @@ class PerceptronBP : public BPredUnit
 
     public:
         struct BPHistory {
-            unsigned globalHistory;
+            uint64_t globalHistory;
             bool globalPredTaken;
         };
